fix: 修复 DeviceController 集成测试 ID 类型 Integer->Long，解决 CI 测试编译失败

+++ src/test/java/top/aiolife/record/api/DeviceControllerIntegrationTest.java
@@ -28,7 +28,7 @@ class DeviceControllerIntegrationTest extends BaseIntegrationTest {
 
     @Test
     void testQuery_查询设备列表() {
-        Integer deviceId = (int) (System.currentTimeMillis() % 1000000 + 900000);
+        Long deviceId = System.currentTimeMillis() % 1000000 + 900000;
         deviceMapper.insert(createDevice(deviceId));
 
         CommonQuery<DeviceEntity> query = new CommonQuery<>();
@@ -42,7 +42,7 @@ class DeviceControllerIntegrationTest extends BaseIntegrationTest {
         assertTrue(response.getData().getTotal() >= 1);
     }
 
-    private DeviceEntity createDevice(Integer deviceId) {
+    private DeviceEntity createDevice(Long deviceId) {
         DeviceEntity entity = new DeviceEntity();
         entity.setId(deviceId);
         entity.setUserId(TEST_USER_ID);
